pyproject.toml
[tool.poetry]
name = "syrobot-1"
version = "0.1.0"
description = ""
authors = ["Glowstick"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
